+++ src/main/java/com/example/namo2/domain/moim/application/converter/MoimScheduleResponseConverter.java
@@ -107,4 +107,5 @@ public class MoimScheduleResponseConverter {
                 .color(moimAndUser.getColor())
                 .build();
     }
+
 }
