+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/ScheduleRepositoryCustom.java
@@ -20,5 +20,5 @@ public interface ScheduleRepositoryCustom {
 
     List<MoimScheduleResponse.MoimScheduleDto> findMonthScheduleInMoim(Long moimId, LocalDateTime startDate, LocalDateTime endDate);
 
-    public List<ScheduleResponse.GetScheduleDto> findMoimSchedulesByUserId(User user, LocalDateTime startDate, LocalDateTime endDate);
+    List<ScheduleResponse.GetScheduleDto> findMoimSchedulesByUserId(User user, LocalDateTime startDate, LocalDateTime endDate);
 }
